requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
pyjwt>=2.8.0
cryptography>=42.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
